Refactored test package to skip package scan
