+++ platform/dolphin-platform-remoting-server/src/test/java/com/canoo/dp/impl/server/controller/repository/TestController.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dp.impl.server.controller;
+package com.canoo.dp.impl.server.controller.repository;
 
+import com.canoo.dp.impl.server.controller.TestBean;
 import com.canoo.platform.remoting.server.DolphinAction;
 import com.canoo.platform.remoting.server.DolphinController;
 import com.canoo.platform.remoting.server.DolphinModel;
